+++ conductor_api/src/interface.rs
@@ -22,7 +22,7 @@ use std::{
     thread,
 };
 
-use conductor::{ConductorAdmin, ConductorTestAdmin, ConductorUiAdmin, CONDUCTOR};
+use conductor::{ConductorAdmin, ConductorDebug, ConductorTestAdmin, ConductorUiAdmin, CONDUCTOR};
 use config::{
     AgentConfiguration, Bridge, DnaConfiguration, InstanceConfiguration, InterfaceConfiguration,
     InterfaceDriver, UiBundleConfiguration, UiInterfaceConfiguration,
@@ -412,6 +412,7 @@ impl ConductorApiBuilder {
     ///     Params:
     ///     * `interface_id`: Which interface to add the instance to?
     ///     * `instance_id`: Which instance to add?
+    ///     * `alias`: (Optional) Local name of the instance within this interface
     ///
     ///  * `admin/interface/remove_instance`
     ///     Remove an instance from a given interface.
@@ -630,7 +631,12 @@ impl ConductorApiBuilder {
                 let params_map = Self::unwrap_params_map(params)?;
                 let interface_id = Self::get_as_string("interface_id", &params_map)?;
                 let instance_id = Self::get_as_string("instance_id", &params_map)?;
-                conductor_call!(|c| c.add_instance_to_interface(&interface_id, &instance_id))?;
+                let alias = Self::get_as_string("alias", &params_map).ok();
+                conductor_call!(|c| c.add_instance_to_interface(
+                    &interface_id,
+                    &instance_id,
+                    &alias
+                ))?;
                 Ok(json!({"success": true}))
             });
 
@@ -710,6 +716,59 @@ impl ConductorApiBuilder {
             let bridges =
                 conductor_call!(|c| Ok(c.config().bridges) as Result<Vec<Bridge>, String>)?;
             Ok(serde_json::to_value(bridges).map_err(|_| jsonrpc_core::Error::internal_error())?)
+        });
+
+        self
+    }
+
+    /// Adds functions useful for debugging.
+    ///
+    /// - `debug/running_instances`
+    ///     Get all currently running instances.
+    ///     Returns an array of instance ID strings.
+    ///
+    /// - `debug/state_dump`
+    ///   Returns a JSON object with all relevant fields of an instance's state.
+    ///   Params:
+    ///   - `instance_id` ID of the instance of which the state is requested
+    ///
+    /// - `debug/fetch_cas`
+    ///   Returns content of a given instance's CAS.
+    ///   Params:
+    ///   - `instance_id` ID of the instance of which's CAS content is requested
+    ///   - `address` Address (hash) of the content that is requests
+    ///   Returns an object of the form: {type:"<entry type>", content: "<content>"}
+    ///
+    pub fn with_debug_functions(mut self) -> Self {
+        self.io
+            .add_method("debug/running_instances", move |_params| {
+                let running_instances_ids = conductor_call!(|c| c.running_instances())?;
+                Ok(serde_json::to_value(running_instances_ids)
+                    .map_err(|_| jsonrpc_core::Error::internal_error())?)
+            });
+
+        self.io.add_method("debug/state_dump", move |params| {
+            let params_map = Self::unwrap_params_map(params)?;
+            let instance_id = Self::get_as_string("instance_id", &params_map)?;
+
+            let dump = conductor_call!(|c| c.state_dump_for_instance(&instance_id))?;
+
+            Ok(serde_json::to_value(dump).map_err(|_| jsonrpc_core::Error::internal_error())?)
+        });
+
+        self.io.add_method("debug/fetch_cas", move |params| {
+            let params_map = Self::unwrap_params_map(params)?;
+            let instance_id = Self::get_as_string("instance_id", &params_map)?;
+            let address = Self::get_as_string("address", &params_map)?;
+
+            let (entry_type, content) = conductor_call!(
+                |c| c.get_type_and_content_from_cas(&Address::from(address), &instance_id)
+            )?;
+
+            Ok(json!({
+                "type": entry_type,
+                "content": content,
+            }))
         });
 
         self