+++ core/src/lib.rs
@@ -43,6 +43,7 @@ pub mod persister;
 pub mod scheduled_jobs;
 pub mod signal;
 pub mod state;
+pub mod state_dump;
 pub mod workflows;
 
 mod conductor_api;